+++ src/OneAppServiceProvider.php
@@ -2,10 +2,12 @@
 
 namespace Envor\OneApp;
 
+use Envor\OneApp\Commands\FolioMakeCommand;
 use Envor\OneApp\Commands\InvitationOnlyCommand;
 use Envor\OneApp\Commands\NavigationCommand;
 use Envor\OneApp\Commands\OneAppCommand;
 use Envor\OneApp\Commands\PassportCommand;
+use Envor\OneApp\Commands\VoltMakeCommand;
 use Spatie\LaravelPackageTools\Package;
 use Spatie\LaravelPackageTools\PackageServiceProvider;
 
@@ -25,6 +27,8 @@ class OneAppServiceProvider extends PackageServiceProvider
                 PassportCommand::class,
                 InvitationOnlyCommand::class,
                 NavigationCommand::class,
+                VoltMakeCommand::class,
+                FolioMakeCommand::class,
             ]);
     }
 }